Enable APOC Full Testing in CI (pr 2891)
* Don't ignore tests which trigger docker exceptions

Tests were ignored if they triggered a docker container exception during start up. We should not be ignoring tests because of this and instead we should be solving the root cause of the issue. Ignoring this tests is a problem because as we've seen in the previous commit they would never run at all since the container wasn't exposing the correct ports and was failing at startup.

Fix Neo4jContainer

The container didn't expose the ports required to be started up, and would therefor never start. We've added fixed ports. In the future it may be possible to add dynamic ports so that we can start multiple containers at the same time, but this would presently require quite a lot of work.

Add Gephi dependency

GephiTest used to not be able to run in the CI because it assumed that the real service was running locally. After much initial investigation into whether it would be possible to either reuse (or make my own) Gephi docker container, I decided not to go down the Docker route. This is because Gephi is quite an old piece of software and does not expose a programmatic way to install the GraphStreaming plugin that we need it to have.

Instead, we did the next best thing which was to create the GephiMock.

Add MySQL dependency

There were some tests which used MySQL that were failing (but ignored). Removing the line which ignored the test made the test fail. The problem was a missing MySQL dependency. We've done:
- Add MySQLContainerExtension which is a testcontainer of MySQL with sensible defaults
- Add MySQLJdbcTest to make sure the testcontainer works
- Fix the PeriodicExtendedTest test

Fix MySQL load test

- The test was fetching countrylanguage and comparing against the head of the list. This was flaky because there are many countrylanguages and they were not always returned in the same order. However these is only one country entity, so it works consistently.

Add GoogleCloudStorage dependency

Like many other tests in APOC Full, this dependency was missing and so we decided to add it in. We deleted some of the tests which won't work:
  - Test which tested that you could access a file at the root level without a bucket. This would never work because it's simply not possible in GCP to have a file without a bucket.
  - Test which tested the auth integration with GCP. I decided to remove this because I felt that it didn't provide that much value with a mock service.

Delete NLP tests which don't use mocks

We have three categories of NLP tests which were all using the live services directly and would of course not work in an isolated environment:
- AzureNLPApiTest
- GCPNLPApiTest
- AWSNLPApiTest

But it turns out that these tests already have implementations which use a dummy client which we can use instead:
- AzureNLPApiTestWithDummyClient
- GCPNLPApiTestWithDummyClient
- AWSNLPApiTestWithDummyClient

Ignore broken tests

These tests are failing for one reason or another, and they need to be fixed in a subsequent PR to not explode the size of the current PR.
These tests are real bugs either in the test file or in the source code.

Run :spotlessApply

Fix BoltTest

It has its own password which was different from the admin password, and was not working because of it.

Remove MetricsTest debugger

Simplify BoltTest.fromLocal

* Add WaitStrategies

- Add for MySQLContainer
- Add for GCSContainer
- Delete old startup strategy for Neo4jContainer
- Tweak the startup time to 120 seconds on Neo4jContainer as it was before

* Remove fixed port numbers

- We removed fixed ports, so now use random ports instead
- The container URLs need to be generated each time
- Generate URLs for GCSContainer
- Generate URL for Neo4jContainer
- Expose Neo4jContainer ports

* Document GephiMock

* Document ignored tests

Add todos or documentation for the ignored tests

* Improve MultiDBTests

These tests are currently quite flaky. Attempt to stabilise:
- Reusing Driver
- Reusing Sessions
- Reusing Transaction

* Log as much as possible

- Always log Neo4jContainer stdout
- Log StartupTest exceptions
- Configure Neo4jContainer to log stdout at debug level, authentication logs, security logs, debug logs, http logs
- Log Neo4jContainer debug.log, http.log, and security.log on startup failure
- Log Neo4jContainer wait strategy attempts